BFS is a no go.
We have been steadily working on this project through the Summer but came to a blocking point this week that we thought we could overcome by the end of today but unfortunately have not been able to see a way forward. Because of this we will not be ready to demo on Saturday for the Seattle Retro Computer Society. However, we do have a way forward from these blockers that will take more time to complete. The way we see it, instead of finishing this week we now foresee finishing this project by the end of next week. Additionally, we will plan to host our own testing/demo day to do a quality check on our AI code. Below I list the blockers that we are currently facing as well the remediation for it:

1. The Atari system can only handle so much memory so for the BFS algorithm that I am running it slows down the entire system to the point where user input can’t be taken in. 

This is based on the design of both the algorithm that I have configured and the game itself. The algorithm that we originally implemented is based on the most basic structures and thus has terrible efficiency. In order to fix this we need to build two structures a priority queue and a hash set. To first include a queue such that moves that are more advantageous can be explored first. Second, to have random access like an array does with efficient searching abilities. Both of these structures will allow us to write BFS in a more efficient manner for the AI tank to find the player's tank. 

2. The current algorithm that I have in play does not account for the fact that the tank is larger than one pixel so ran into issues on that front too. 
Originally, while I was writing the code for the game. I made an incorrect assumption that each tank sprite is only one pixel large. This caused a big issue because when the AI tank wanted to fire at the player's tank because the missle doesn't come out of the top pixel of the tank itself. Instead it comes out of the middle. This fix again is about a days worth of work but we just wanted to let you know that we will not be able to meet the deadline for demo-ing it at the SRCS @ Kirkland.

We will continue to work more on this project this week and will provide you updates as we complete this project.